fix(architecture): move base EF configurations to Configurations/ folder
- Move LegalAgreementBaseConfiguration from Internal/ to Configurations/
  and update namespace to Granit.Privacy.EntityFrameworkCore.Configurations
- Move VersionedWorkflowEntityBaseConfiguration from root to Configurations/
  and update namespace to Granit.Workflow.EntityFrameworkCore.Configurations

Fixes two architecture test failures:
- Namespace_should_match_folder_structure_in_src
- EfCore_configurations_should_reside_in_Configurations_folder
